vue colada - rename files
