+++ packages/vue-colada/src/router-utils.ts
@@ -1,6 +1,6 @@
 import type { Client, NestedClient } from '@orpc/contract'
-import { createGeneralUtils, type GeneralUtils } from './utils-general'
-import { createProcedureUtils, type ProcedureUtils } from './utils-procedure'
+import { createGeneralUtils, type GeneralUtils } from './general-utils'
+import { createProcedureUtils, type ProcedureUtils } from './procedure-utils'
 
 export type RouterUtils<T extends NestedClient<any>> =
   T extends Client<infer UClientContext, infer UInput, infer UOutput, infer UError>
